Move parameter-ref based exec to submodule
