+++ src/pdag/examples/diamond_mdp.py
@@ -93,7 +93,7 @@ class DiamondMdpModel(pdag.Model):
 if __name__ == "__main__":
     from rich import print  # noqa: A004
 
-    from pdag._exec import exec_core_model
+    from pdag._exec import exec_core_model_via_paramref as exec_core_model
 
     core_model = DiamondMdpModel.to_core_model()
     print(core_model)
